+++ docs/changelog.rst
@@ -79,6 +79,8 @@ Detailed list of changes
 
 - Linux: Fix a regression in 0.36.0 that caused font features defined via fontconfig to be ignored (:iss:`7773`)
 
+- :ac:`goto_tab`: Allow numbers less than ``-1`` to go to the Nth previously active tab
+
 0.36.1 [2024-08-24]
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~